refactor: repositoryファイルのディレクトリ構造を修正
- ArticleRepositoryImpl.ts を infrastructure/external/ に移動
- BookmarkRepositoryImpl.ts を infrastructure/persistence/ に移動
- 各use-caseのimportパスを新しいディレクトリ構造に更新
- FetchArticlesUseCase.tsの関数形式を修正
