+++ graphql/src/application/use-cases/articles/FetchArticlesUseCase.ts
@@ -1,6 +1,8 @@
 import type { Article } from "../../../domain/articles/Article";
-import * as articleRepository from "../../../infrastructure/repositories/ArticleRepositoryImpl";
+import * as articleRepository from "../../../infrastructure/external/ArticleRepositoryImpl";
 
-export const execute = async (page: number): Promise<Article[]> => {
+export const fetchArticlesUseCase = async (
+  page: number,
+): Promise<Article[]> => {
   return articleRepository.fetchArticles(page);
 };
