+++ graphql/src/application/use-cases/bookmarks/CreateBookmarkUseCase.ts
@@ -2,9 +2,9 @@ import type {
   Bookmark,
   CreateBookmarkInput,
 } from "../../../domain/bookmarks/Bookmark";
-import * as bookmarkRepository from "../../../infrastructure/repositories/BookmarkRepositoryImpl";
+import * as bookmarkRepository from "../../../infrastructure/persistence/BookmarkRepositoryImpl";
 
-export const execute = async (
+export const createBookmarkUseCase = async (
   input: CreateBookmarkInput,
 ): Promise<Bookmark> => {
   return bookmarkRepository.create(input);
