+++ graphql/src/application/use-cases/bookmarks/DeleteBookmarkUseCase.ts
@@ -1,5 +1,5 @@
-import * as bookmarkRepository from "../../../infrastructure/repositories/BookmarkRepositoryImpl";
+import * as bookmarkRepository from "../../../infrastructure/persistence/BookmarkRepositoryImpl";
 
-export const execute = async (id: string): Promise<boolean> => {
+export const deleteBookmarkUseCase = async (id: string): Promise<boolean> => {
   return bookmarkRepository.deleteBookmark(id);
 };
