+++ graphql/src/application/use-cases/bookmarks/FetchBookmarkByIdUseCase.ts
@@ -1,6 +1,8 @@
 import type { Bookmark } from "../../../domain/bookmarks/Bookmark";
-import * as bookmarkRepository from "../../../infrastructure/repositories/BookmarkRepositoryImpl";
+import * as bookmarkRepository from "../../../infrastructure/persistence/BookmarkRepositoryImpl";
 
-export const execute = async (id: string): Promise<Bookmark | null> => {
+export const fetchBookmarkByIdUseCase = async (
+  id: string,
+): Promise<Bookmark | null> => {
   return bookmarkRepository.findById(id);
 };
