+++ graphql/src/application/use-cases/bookmarks/UpdateBookmarkUseCase.ts
@@ -2,9 +2,9 @@ import type {
   Bookmark,
   UpdateBookmarkInput,
 } from "../../../domain/bookmarks/Bookmark";
-import * as bookmarkRepository from "../../../infrastructure/repositories/BookmarkRepositoryImpl";
+import * as bookmarkRepository from "../../../infrastructure/persistence/BookmarkRepositoryImpl";
 
-export const execute = async (
+export const updateBookmarkUseCase = async (
   id: string,
   input: UpdateBookmarkInput,
 ): Promise<Bookmark | null> => {
