--- graphql/src/domain/articles/ArticleRepository.ts
@@ -1,5 +1,0 @@
-import type { Article } from "./Article";
-
-export type ArticleRepository = {
-  fetchArticles(page: number): Promise<Article[]>;
-};
