+++ graphql/src/infrastructure/persistence/BookmarkRepositoryImpl.ts
